requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
numpy==1.26.4
psycopg2-binary==2.9.9
scikit-learn==1.4.0
streamlit==1.31.1
python-dotenv==1.0.0
minio==7.2.0
